back/requirements.txt

gunicorn
Flask
MindustryCompiler
